Add files via upload
The Tribute Page: Marie Curie provides a brief overview of a webpage created to honor Marie Curie. It highlights her scientific achievements in radioactivity and Nobel Prize awards in physics and chemistry. The README details key features like a responsive design, a centered layout, and a "Learn More" link directing users to Curie's Wikipedia page. It also lists HTML and CSS as the technologies used and includes instructions on how to open the tribute page. This project serves as a basic practice in HTML and CSS while celebrating an influential scientist.